Multitenant payment processor configuration
