+++ Makefile
@@ -1,4 +1,5 @@
 NODE_BIN=./node_modules/.bin
+DIFF_COVER_BASE_BRANCH=master
 
 help:
 	@echo '                                                                                     		'
@@ -77,6 +78,12 @@ static:
 
 html_coverage:
 	coverage html && open htmlcov/index.html
+
+diff_coverage: validate fast_diff_coverage
+
+fast_diff_coverage:
+	coverage xml
+	diff-cover coverage.xml --compare-branch=$(DIFF_COVER_BASE_BRANCH)
 
 accept:
 	nosetests --with-ignore-docstrings -v acceptance_tests
