+++ ecommerce/core/tests/test_models.py
@@ -1,5 +1,20 @@
-from ecommerce.core.models import BusinessClient, User
+import ddt
+import mock
+from django.conf import settings
+from django.contrib.sites.models import Site
+from django.core.exceptions import ValidationError
+from django.test import override_settings
+
+from ecommerce.core.models import BusinessClient, User, SiteConfiguration
+from ecommerce.core.tests import toggle_switch
+from ecommerce.extensions.payment.tests.processors import DummyProcessor, AnotherDummyProcessor
 from ecommerce.tests.testcases import TestCase
+
+
+def _make_site_config(payment_processors_str, site_id=1):
+    site = Site.objects.get(id=site_id)
+
+    return SiteConfiguration(site=site, payment_processors=payment_processors_str)
 
 
 class UserTests(TestCase):
@@ -45,3 +60,124 @@ class BusinessClientTests(TestCase):
     def test_str(self):
         client = BusinessClient.objects.create(name='TestClient')
         self.assertEquals(str(client), 'TestClient')
+
+
+@ddt.ddt
+class SiteConfigurationTests(TestCase):
+    @ddt.data(
+        ("paypal", {"paypal"}),
+        ("paypal ", {"paypal"}),
+        ("paypal,cybersource", {"paypal", "cybersource"}),
+        ("paypal, cybersource", {"paypal", "cybersource"}),
+        ("paypal,cybersource,something_else", {"paypal", "cybersource", "something_else"}),
+        ("paypal , cybersource , something_else", {"paypal", "cybersource", "something_else"}),
+    )
+    @ddt.unpack
+    def test_payment_processor_field_parsing(self, payment_processors_str, expected_result):
+        """
+        Tests that comma-separated payment processor string is correctly converted to a set of payment processor names
+        :param str payment_processors_str: comma-separated string of processor names (potentially with spaces)
+        :param set[str] expected_result: expected payment_processors_set result
+        """
+        site_config = _make_site_config(payment_processors_str)
+        self.assertEqual(site_config.payment_processors_set, expected_result)
+
+    @ddt.data("paypal", "paypal, cybersource", "paypal , cybersource")
+    def test_clean_fields_valid_values_pass_validation(self, payment_processors_str):
+        """
+        Tests that valid payment_processors value passes validation
+        :param str payment_processors_str: comma-separated string of processor names (potentially with spaces)
+        """
+        site_config = _make_site_config(payment_processors_str)
+        with mock.patch("ecommerce.extensions.payment.helpers.get_processor_class_by_name") as patched_proc_by_name:
+            patched_proc_by_name.return_value = DummyProcessor
+            try:
+                site_config.clean_fields()
+            except ValidationError as exc:
+                self.fail(exc.message)
+
+    @ddt.data(" ", "  \t ", "\t\n\r")
+    def test_clean_fields_whitespace_payment_processor_fail_validation(self, payment_processors_str):
+        """
+        Tests that whitespace-only payment_processor values fail validation
+        :param str payment_processors_str: comma-separated string of processor names (potentially with spaces)
+        """
+        site_config = _make_site_config(payment_processors_str)
+        with self.assertRaises(ValidationError) as err:
+            site_config.clean_fields()
+            self.assertEqual(
+                err.message, "Invalid payment processors field: must not only contain whitespace characters"
+            )
+
+    def test_clean_fields_unknown_payment_processor_fail_validation(self):
+        """
+        Tests that  validation fails if payment_processors field contains unknown payment processor names
+        """
+        site_config = _make_site_config("unknown_payment_processor")
+
+        with self.assertRaises(ValidationError):
+            site_config.clean_fields()
+
+    def test_clean_fields_payment_processor_excluded_always_pass(self):
+        """
+        Tests that `clean_fields` pass if "payment_processors" are excluded, regardless of validity
+        """
+        site_config = _make_site_config("")
+        site_config.clean_fields(exclude={"payment_processors"})
+
+        site_config.payment_processors = "irrelevant-get_processor_by_name-is-patched"
+        site_config.clean_fields(exclude={"payment_processors"})
+
+    @staticmethod
+    def _enable_processor_switches(processors):
+        for processor in processors:
+            toggle_switch(settings.PAYMENT_PROCESSOR_SWITCH_PREFIX + processor.NAME, True)
+
+    @override_settings(PAYMENT_PROCESSORS=[
+        'ecommerce.extensions.payment.tests.processors.DummyProcessor',
+        'ecommerce.extensions.payment.tests.processors.AnotherDummyProcessor',
+    ])
+    @ddt.data(
+        ([], []),
+        ([DummyProcessor], [DummyProcessor]),
+        ([DummyProcessor, AnotherDummyProcessor], [DummyProcessor, AnotherDummyProcessor]),
+    )
+    @ddt.unpack
+    def test_get_payment_processors(self, processors, expected_result):
+        """ Tests that get_payment_processors returs correct payment processor classes """
+        self._enable_processor_switches(processors)
+        site_config = _make_site_config(",".join(proc.NAME for proc in processors))
+
+        result = site_config.get_payment_processors()
+        self.assertEqual(result, expected_result)
+
+    @override_settings(PAYMENT_PROCESSORS=[
+        'ecommerce.extensions.payment.tests.processors.DummyProcessor',
+    ])
+    def test_get_payment_processors_logs_warning_for_unknown_processors(self):
+        """ Tests that get_payment_processors logs warnings if unknown payment processor codes are seen """
+        processors = [DummyProcessor, AnotherDummyProcessor]
+        site_config = _make_site_config(",".join(proc.NAME for proc in processors))
+        with mock.patch("ecommerce.core.models.log") as patched_log:
+            site_config.get_payment_processors()
+            patched_log.warning.assert_called_once_with(
+                'Unknown payment processors [%s] are configured for site %s',
+                AnotherDummyProcessor.NAME,
+                site_config.site.id
+            )
+
+    @override_settings(PAYMENT_PROCESSORS=[
+        'ecommerce.extensions.payment.tests.processors.DummyProcessor',
+        'ecommerce.extensions.payment.tests.processors.AnotherDummyProcessor',
+    ])
+    @ddt.data(
+        [DummyProcessor],
+        [DummyProcessor, AnotherDummyProcessor]
+    )
+    def test_get_payment_processors_switch_disabled(self, processors):
+        """ Tests that get_payment_processors respects waffle switches """
+        expected_result = []
+        site_config = _make_site_config(",".join(proc.NAME for proc in processors))
+
+        result = site_config.get_payment_processors()
+        self.assertEqual(result, expected_result)
