+++ ecommerce/extensions/api/v2/tests/views/test_payments.py
@@ -6,6 +6,7 @@ from django.core.urlresolvers import reverse
 from django.test import override_settings
 from waffle.models import Switch
 
+from ecommerce.core.models import SiteConfiguration
 from ecommerce.extensions.payment.tests.processors import DummyProcessor, AnotherDummyProcessor
 from ecommerce.tests.testcases import TestCase
 
@@ -16,8 +17,22 @@ class PaymentProcessorListViewTests(TestCase):
     def setUp(self):
         super(PaymentProcessorListViewTests, self).setUp()
         self.token = self.generate_jwt_token_header(self.create_user())
-        self.toggle_payment_processor('dummy', True)
-        self.toggle_payment_processor('another-dummy', True)
+        self.toggle_payment_processor(DummyProcessor.NAME, True)
+        self.toggle_payment_processor(AnotherDummyProcessor.NAME, True)
+
+        site_config, __ = SiteConfiguration.objects.get_or_create(site__id=1)
+
+        old_payment_processors = site_config.payment_processors
+        site_config.payment_processors = ",".join([DummyProcessor.NAME, AnotherDummyProcessor.NAME])
+        site_config.save()
+
+        def reset_site_config():
+            """ Reset method - resets site_config to pre-test state """
+            site_config.payment_processors = old_payment_processors
+            site_config.save()
+
+        self.addCleanup(reset_site_config)
+
         # Clear the view cache
         cache.clear()
 
@@ -55,7 +70,8 @@ class PaymentProcessorListViewTests(TestCase):
         'ecommerce.extensions.payment.tests.processors.DummyProcessor',
     ])
     def test_processor_disabled(self):
-        self.toggle_payment_processor('dummy', False)
+        """  Tests that disabloing payment processor works """
+        self.toggle_payment_processor(DummyProcessor.NAME, False)
         self.assert_processor_list_matches([])
 
     @override_settings(PAYMENT_PROCESSORS=[
@@ -63,6 +79,7 @@ class PaymentProcessorListViewTests(TestCase):
         'ecommerce.extensions.payment.tests.processors.AnotherDummyProcessor',
     ])
     def test_waffle_switches_clear_cache(self):
+        """ Tests that adding a new Switch resets processor cache """
         self.assert_processor_list_matches([DummyProcessor.NAME, AnotherDummyProcessor.NAME])
-        self.toggle_payment_processor('dummy', False)
+        self.toggle_payment_processor(DummyProcessor.NAME, False)
         self.assert_processor_list_matches([AnotherDummyProcessor.NAME])
