+++ ecommerce/extensions/api/v2/views/payments.py
@@ -1,11 +1,9 @@
 """HTTP endpoints for interacting with payments."""
-from django.conf import settings
 from rest_framework import generics
 from rest_framework.permissions import IsAuthenticated
 from rest_framework_extensions.cache.decorators import cache_response
 
 from ecommerce.extensions.api import serializers
-from ecommerce.extensions.payment.helpers import get_processor_class
 
 
 PAYMENT_PROCESSOR_CACHE_KEY = 'PAYMENT_PROCESSOR_LIST'
@@ -28,5 +26,4 @@ class PaymentProcessorListView(generics.ListAPIView):
 
     def get_queryset(self):
         """Fetch the list of payment processor classes based on Django settings."""
-        processors = (get_processor_class(path) for path in settings.PAYMENT_PROCESSORS)
-        return [processor for processor in processors if processor.is_enabled()]
+        return self.request.site.siteconfiguration.get_payment_processors()
