+++ ecommerce/extensions/payment/processors/paypal.py
@@ -4,6 +4,7 @@ import logging
 from urlparse import urljoin
 
 from django.core.urlresolvers import reverse
+from django.utils.functional import cached_property
 from oscar.apps.payment.exceptions import GatewayError
 from oscar.core.loading import get_model
 import paypalrestsdk
@@ -45,6 +46,18 @@ class Paypal(BasePaymentProcessor):
         """
         # Number of times payment execution is retried after failure.
         self.retry_attempts = self.configuration.get('retry_attempts', 1)
+
+    @cached_property
+    def paypal_api(self):
+        """
+        Returns Paypal API instance with appropriate configuration
+        Returns: Paypal API instance
+        """
+        return paypalrestsdk.Api({
+            'mode': self.configuration['mode'],
+            'client_id': self.configuration['client_id'],
+            'client_secret': self.configuration['client_secret']
+        })
 
     @property
     def receipt_url(self):
@@ -113,7 +126,7 @@ class Paypal(BasePaymentProcessor):
         except PaypalWebProfile.DoesNotExist:
             pass
 
-        payment = paypalrestsdk.Payment(data)
+        payment = paypalrestsdk.Payment(data, api=self.paypal_api)
         payment.create()
 
         # Raise an exception for payments that were not successfully created. Consuming code is
@@ -180,7 +193,7 @@ class Paypal(BasePaymentProcessor):
 
         for attempt_count in range(1, available_attempts + 1):
 
-            payment = paypalrestsdk.Payment.find(response.get('paymentId'))
+            payment = paypalrestsdk.Payment.find(response.get('paymentId'), api=self.paypal_api)
             payment.execute(data)
 
             if payment.success():
@@ -267,7 +280,7 @@ class Paypal(BasePaymentProcessor):
         order = source.order
 
         try:
-            payment = paypalrestsdk.Payment.find(source.reference)
+            payment = paypalrestsdk.Payment.find(source.reference, api=self.paypal_api)
             sale = self._get_payment_sale(payment)
 
             if not sale:
