+++ ecommerce/extensions/payment/views.py
@@ -35,7 +35,9 @@ PaymentProcessorResponse = get_model('payment', 'PaymentProcessorResponse')
 
 class CybersourceNotifyView(EdxOrderPlacementMixin, View):
     """ Validates a response from CyberSource and processes the associated basket/order appropriately. """
-    payment_processor = Cybersource()
+    @property
+    def payment_processor(self):
+        return Cybersource()
 
     # Disable atomicity for the view. Otherwise, we'd be unable to commit to the database
     # until the request had concluded; Django will refuse to commit when an atomic() block
@@ -171,7 +173,9 @@ class CybersourceNotifyView(EdxOrderPlacementMixin, View):
 
 class PaypalPaymentExecutionView(EdxOrderPlacementMixin, View):
     """Execute an approved PayPal payment and place an order for paid products as appropriate."""
-    payment_processor = Paypal()
+    @property
+    def payment_processor(self):
+        return Paypal()
 
     # Disable atomicity for the view. Otherwise, we'd be unable to commit to the database
     # until the request had concluded; Django will refuse to commit when an atomic() block
