+++ ecommerce/settings/_oscar.py
@@ -107,23 +107,25 @@ PAYMENT_PROCESSOR_CANCEL_PATH = '/commerce/checkout/cancel/'
 PAYMENT_PROCESSOR_ERROR_PATH = '/commerce/checkout/error/'
 
 PAYMENT_PROCESSOR_CONFIG = {
-    'cybersource': {
-        'profile_id': None,
-        'access_key': None,
-        'secret_key': None,
-        'payment_page_url': None,
-        'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
-        'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
-        'send_level_2_3_details': True,
-    },
-    'paypal': {
-        # 'mode' can be either 'sandbox' or 'live'
-        'mode': None,
-        'client_id': None,
-        'client_secret': None,
-        'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
-        'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
-        'error_path': PAYMENT_PROCESSOR_ERROR_PATH,
+    'edx': {
+        'cybersource': {
+            'profile_id': None,
+            'access_key': None,
+            'secret_key': None,
+            'payment_page_url': None,
+            'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
+            'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'send_level_2_3_details': True,
+        },
+        'paypal': {
+            # 'mode' can be either 'sandbox' or 'live'
+            'mode': None,
+            'client_id': None,
+            'client_secret': None,
+            'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
+            'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'error_path': PAYMENT_PROCESSOR_ERROR_PATH,
+        },
     },
 }
