+++ ecommerce/settings/test.py
@@ -66,27 +66,54 @@ EDX_API_KEY = 'replace-me'
 
 # PAYMENT PROCESSING
 PAYMENT_PROCESSOR_CONFIG = {
-    'cybersource': {
-        'soap_api_url': 'https://ics2wstest.ic3.com/commerce/1.x/transactionProcessor/CyberSourceTransaction_1.115.wsdl',
-        'merchant_id': 'fake-merchant-id',
-        'transaction_key': 'fake-transaction-key',
-        'profile_id': 'fake-profile-id',
-        'access_key': 'fake-access-key',
-        'secret_key': 'fake-secret-key',
-        'payment_page_url': 'https://replace-me/',
-        'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
-        'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
-        'send_level_2_3_details': True,
+    'edx': {
+        'cybersource': {
+            'soap_api_url': 'https://ics2wstest.ic3.com/commerce/1.x/transactionProcessor/CyberSourceTransaction_1.115.wsdl',
+            'merchant_id': 'fake-merchant-id',
+            'transaction_key': 'fake-transaction-key',
+            'profile_id': 'fake-profile-id',
+            'access_key': 'fake-access-key',
+            'secret_key': 'fake-secret-key',
+            'payment_page_url': 'https://replace-me/',
+            'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
+            'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'send_level_2_3_details': True,
+        },
+        'paypal': {
+            'mode': 'sandbox',
+            'client_id': 'fake-client-id',
+            'client_secret': 'fake-client-secret',
+            'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
+            'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'error_path': PAYMENT_PROCESSOR_ERROR_PATH,
+        },
+        'invoice': {}
     },
-    'paypal': {
-        'mode': 'sandbox',
-        'client_id': 'fake-client-id',
-        'client_secret': 'fake-client-secret',
-        'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
-        'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
-        'error_path': PAYMENT_PROCESSOR_ERROR_PATH,
-    },
+    'other': {
+        'cybersource': {
+            'soap_api_url': 'https://ics2wstest.ic3.com/commerce/1.x/transactionProcessor/CyberSourceTransaction_1.115.wsdl',
+            'merchant_id': 'other-fake-merchant-id',
+            'transaction_key': 'other-fake-transaction-key',
+            'profile_id': 'other-fake-profile-id',
+            'access_key': 'other-fake-access-key',
+            'secret_key': 'other-fake-secret-key',
+            'payment_page_url': 'https://replace-me/',
+            'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
+            'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'send_level_2_3_details': True,
+        },
+        'paypal': {
+            'mode': 'sandbox',
+            'client_id': 'pther-fake-client-id',
+            'client_secret': 'pther-fake-client-secret',
+            'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
+            'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'error_path': PAYMENT_PROCESSOR_ERROR_PATH,
+        },
+        'invoice': {}
+    }
 }
+
 # END PAYMENT PROCESSING
