+++ requirements/test.txt
@@ -13,3 +13,4 @@ pep8==1.6.2
 pylint==1.5.0
 selenium>=2.53.1
 testfixtures==4.5.0
+diff-cover==0.9.6
